% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_journal_detail.R
\name{get_journal_detail}
\alias{get_journal_detail}
\title{Get Journal Detail from TR Dizin}
\usage{
get_journal_detail(id)
}
\arguments{
\item{id}{A character string or numeric value representing the journal's ID.}
}
\value{
A parsed list containing the JSON response for the specified journal.
}
\description{
This function retrieves detailed information about a specific journal
from the TR Dizin database using the journal's numeric ID.
}
