% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publication_detail.R
\name{get_publication_detail}
\alias{get_publication_detail}
\title{Get Publication Detail from TR Dizin}
\usage{
get_publication_detail(id)
}
\arguments{
\item{id}{A character string or numeric value representing the publication's ID.}
}
\value{
A parsed list containing the JSON response for the specified publication.
}
\description{
This function retrieves detailed information about a specific publication
from the TR Dizin database using the publication's unique ID.
}
